pyproject.toml
#:schema https://json.schemastore.org/pyproject.json
# https://packaging.python.org/en/latest/specifications/pyproject-toml/

[build-system]
build-backend = "hatchling.build"
requires = ["hatchling"]

[dependency-groups]
dev = ["icecream~=2.1.4"]

[project]
authors = [
  { email = "30631553+liblaf@users.noreply.github.com", name = "liblaf" },
]
dependencies = [
  "gitpython~=3.1.44",
  "glom~=24.11.0",
  "lazy-loader~=0.4",
  "liblaf-avocado~=0.1.0",
  "loguru~=0.7.3",
  "pydantic-settings~=2.7.1",
]
description = "Add your description here"
keywords = []
license = "MIT"
name = "liblaf-banana"
readme = "docs/README.md"
requires-python = "~=3.12"
version = "0.1.0"

[project.optional-dependencies]
comet = ["comet-ml~=3.47.6"]

[project.urls]
"Changelog" = "https://github.com/liblaf/banana/blob/main/CHANGELOG.md"
"Homepage" = "https://github.com/liblaf/banana"
"Issue Tracker" = "https://github.com/liblaf/banana/issues"
"Release Notes" = "https://github.com/liblaf/banana/releases"
"Source Code" = "https://github.com/liblaf/banana"

[tool.check-wheel-contents]
ignore = ["W002"]

[tool.hatch.build.targets.wheel]
packages = ["src/banana"]
